+++ domain/domain-pos/src/main/java/domain/pos/receipt/entity/TableWithNonAdjustReceipt.java
@@ -25,11 +25,11 @@ public class TableWithNonAdjustReceipt {
 
 	public static TableWithNonAdjustReceipt of(Table table, List<Receipt> receipts) {
 		Receipt tableReceipt = receipts.stream()
-			.filter(receipt -> receipt.getTable().getTableId().equals(table.getTableId()))
+			.filter(receipt -> receipt.getTable().getId().equals(table.getId()))
 			.findFirst()
 			.orElse(null);
 		return TableWithNonAdjustReceipt.builder()
-			.tableId(table.getTableId())
+			.tableId(table.getId())
 			.tableNumber(table.getTableNumber())
 			.isActive(table.getIsActive())
 			.nonAdjustReceipt(tableReceipt)
